# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a trial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Trial booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking confirmation
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/signup:
    post:
      operationId: signUp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Signed up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400": { description: Invalid signup, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401": { description: Invalid credentials, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401": { description: Authentication required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
  /parent/kids:
    get:
      operationId: listParentKids
      tags: [parent]
      responses:
        "200": { description: Kids, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/KidProfile" } } } } }
    post:
      operationId: createParentKid
      tags: [parent]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/KidProfileInput" } }
      responses:
        "201": { description: Kid, content: { application/json: { schema: { $ref: "#/components/schemas/KidProfile" } } } }
...
  /analytics:
    get:
      operationId: getAnalytics
      tags: [analytics]
      responses:
        "200": { description: Deprecated employee analytics, content: { application/json: { schema: { $ref: "#/components/schemas/OpsAnalytics" } } } }
        "403": { description: Employee role required }
  /ops:
    get:
      operationId: getOpsAnalytics
      tags: [ops]
      responses:
        "200": { description: Operations analytics, content: { application/json: { schema: { $ref: "#/components/schemas/OpsAnalytics" } } } }
        "403": { description: Employee role required }
components:
  schemas:
...
    KidProfile:
      type: object
      required: [id, name, age]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        age: { type: integer, minimum: 4 }
    KidProfileInput:
      type: object
      required: [name, age]
      properties:
        name: { type: string, minLength: 1 }
        age: { type: integer, minimum: 4 }
    AnalyticsSummary:
      type: object
      required: [periodLabel, GMV, dabbleRevenue, totalBookings, pendingBookings, activeCoaches, searches, conversion, avgBookingValue, newParentSignups, societiesLive, bookingsByCategory, bookingsByArea, bookingsOverTime, topCoaches, recentBookings]
      properties:
        periodLabel: { type: string }
        bookingCount: { type: integer }
        GMV: { type: integer }
        dabbleRevenue: { type: integer }
        totalBookings: { type: integer }
        pendingBookings: { type: integer }
        activeCoaches: { type: integer }
        searches: { type: integer }
        conversion: { type: number }
        avgBookingValue: { type: number }
        newParentSignups: { type: integer }
        societiesLive: { type: integer }
        bookingsByCategory: { type: array, items: { $ref: "#/components/schemas/AnalyticsSeriesPoint" } }
        bookingsByArea: { type: array, items: { $ref: "#/components/schemas/AnalyticsSeriesPoint" } }
        bookingsOverTime: { type: array, items: { $ref: "#/components/schemas/AnalyticsSeriesPoint" } }
        topCoaches: { type: array, items: { $ref: "#/components/schemas/TopCoach" } }
        recentBookings: { type: array, items: { $ref: "#/components/schemas/RecentBooking" } }
    OpsAnalytics:
      allOf:
        - $ref: "#/components/schemas/AnalyticsSummary"
    AnalyticsSeriesPoint:
      type: object
      required: [label, bookings, GMV]
      properties:
        label: { type: string }
        bookings: { type: integer }
        GMV: { type: integer }
    RecentBooking:
      type: object
      required: [parent, child, activity, coach, amount, status, time]
      properties:
        parent: { type: string }
        child: { type: string }
        activity: { type: string }
        coach: { type: string }
        amount: { type: integer }
        status: { type: string, enum: [paid, pending] }
        time: { type: string }
    TopCoach:
      type: object
      required: [coachId, coachName, bookings, seatsBooked, GMV, coachEarnings, revenue]
      properties:
        coachId: { type: string }
        coachName: { type: string }
        bookings: { type: integer }
        seatsBooked: { type: integer }
        GMV: { type: integer }
        coachEarnings: { type: integer }
        revenue: { type: integer }
...
    AuthUser:
      type: object
      required: [id, email, name, role, linkedCoachId]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [parent, coach, employee] }
        linkedCoachId: { type: ["string", "null"] }
    SignupInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        role: { type: string, enum: [parent, coach] }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
...
    Coach:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        activity:
          type: string
        venue:
          type: string
        area:
          type: string
        distance:
          type: string
        rating:
          type: number
        reviews:
          type: integer
        price:
          type: integer
        priceLabel:
          type: string
        ageRange:
          type: string
        experience:
          type: string
        description:
          type: string
        verified:
          type: boolean
        trialAvailable:
          type: boolean
        imageUrl:
          type: string
        accent:
          type: string
        highlights:
          type: array
          items:
            type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TrialSlot"
        sessionFormats:
          type: array
          items:
            type: string
        venueType:
          type: string
        serviceAreas:
          type: array
          items:
            type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        parentAccompanied:
          type: boolean
      required:
        - id
        - name
        - activity
        - venue
        - area
        - distance
        - rating
        - reviews
        - price
        - priceLabel
        - ageRange
        - experience
        - description
        - verified
        - trialAvailable
        - imageUrl
        - accent
        - highlights
        - slots
        - sessionFormats
        - venueType
        - serviceAreas
        - ageMin
        - ageMax
        - parentAccompanied
    TrialSlot:
      type: object
      properties:
        id:
          type: string
        day:
          type: string
        date:
          type: string
        time:
          type: string
        label:
          type: string
      required:
        - id
        - day
        - date
        - time
        - label
    BookingInput:
      type: object
      properties:
        coachId:
          type: string
        slotId:
          type: string
        childName:
          type: string
        childAge:
          type: integer
          minimum: 4
          maximum: 18
        parentName:
          type: string
        parentEmail:
          type: string
        parentPhone:
          type: string
        seats:
          type: integer
          minimum: 1
          maximum: 10
      required:
        - coachId
        - slotId
        - childName
        - childAge
        - parentName
        - parentEmail
        - parentPhone
...
    Booking:
      allOf:
        - $ref: "#/components/schemas/BookingInput"
        - type: object
          properties:
            id:
              type: string
            coach:
              $ref: "#/components/schemas/Coach"
            slot:
              $ref: "#/components/schemas/TrialSlot"
            trialFee:
              type: integer
            serviceFee:
              type: integer
            coachFee:
              type: integer
            dabbleFee:
              type: integer
            seats:
              type: integer
              minimum: 1
              maximum: 10
            total:
              type: integer
            createdAt:
              type: string
          required:
            - id
            - coach
            - slot
            - trialFee
            - serviceFee
             - coachFee
             - dabbleFee
             - seats
            - total
            - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...