# lib/api-zod/src/generated/types/analyticsSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalyticsSeriesPoint } from './analyticsSeriesPoint';
import type { RecentBooking } from './recentBooking';
import type { TopCoach } from './topCoach';

export interface AnalyticsSummary {
  periodLabel: string;
  bookingCount?: number;
  GMV: number;
  dabbleRevenue: number;
  totalBookings: number;
  pendingBookings: number;
  activeCoaches: number;
  searches: number;
  conversion: number;
  avgBookingValue: number;
  newParentSignups: number;
  societiesLive: number;
  bookingsByCategory: AnalyticsSeriesPoint[];
  bookingsByArea: AnalyticsSeriesPoint[];
  bookingsOverTime: AnalyticsSeriesPoint[];
  topCoaches: TopCoach[];
  recentBookings: RecentBooking[];
}