# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  email: string;
  name: string;
  role: AuthUserRole;
  /** @nullable */
  linkedCoachId: string | null;
}