# lib/api-zod/src/generated/types/kidProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KidProfile {
  id: string;
  name: string;
  /** @minimum 4 */
  age: number;
}