# lib/api-zod/src/generated/types/recentBooking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecentBookingStatus } from './recentBookingStatus';

export interface RecentBooking {
  parent: string;
  child: string;
  activity: string;
  coach: string;
  amount: number;
  status: RecentBookingStatus;
  time: string;
}